2.35

+++ app/src/main/java/com/dayzeeco/dayzee/model/Price.kt
@@ -6,6 +6,6 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Price (
-    @SerializedName("value") var price: Int,
+    @SerializedName("value") var price: Double,
     @SerializedName("currency") var currency: String
 ) : Parcelable
